Circumvents issues with apache's PID file, which will show up if the container is stopped and started again.
The Docker daemon sends the SIGTERM (or SIGWINCH) signal to the
process inside the container, but Apache is never cleanly shut
down.

The issue seems to be apachectl, which receives the signal because
it has PID 1 inside the container, but it neither handles nor
propagates the message to its child processes. As the process
ignores the message that the Docker daemon had sent, it will
eventually be killed by a SIGKILL after the timeout expires, leading
to an "unclean" shutdown.

The correct way to handle is would be to handle both SIGTERM and
SIGWINCH correctly, which may be done in the future.